+++ src/SharpX/Extensions/StringExtensions.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Text.RegularExpressions;
 
-namespace SharpX
+namespace SharpX.Extensions
 {
     public static class StringUtil
     {